requirements.txt
astroid==2.4.2
autopep8==1.5.5
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
nltk==3.5
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
rope==0.18.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
symspellpy==6.7.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.58.0
Werkzeug==2.0.1
wordcloud==1.8.1
wrapt==1.12.1
